Unsigned type for self-wrapping counter
The unsigned data type behaves like a hybrid between an integer and a std_logic_vector. You can perform arithmetics on a signal of unsigned type, while at the same time being able to access individual bits within it. Unlike an integer, the unsigned type will wrap silently when used as a counter.