Function and procedure used for avoiding repetitive code
VHDL has two types of subprograms: the function and the procedure. Both are good for avoiding repetitive code in your VHDL project.

The most obvious difference is that the function always returns a value, while the procedure doesn't have a return value. Though, the procedure can return values through the parameter list, while a function cannot.

A function's parameters are always constants, while a procedure can have constants, variables, or signals as parameters. A function only has inputs parameters, while a procedure can have parameters with direction "in", "out", or "inout".

Also, procedures can contain wait-statements, but a wait-statement in a function is a compilation error.